requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.3.0
Flask==1.1.1
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.8
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
swagger-ui-bundle==0.0.5
urllib3==1.25.3
Werkzeug==0.15.5
dnspython==1.16.0
pytest